queryObservers -> subscribeTo*Query*
