+++ src/observation/subscribeToQuery.d.ts
