+++ src/observation/subscribeToQueryReloading/index.d.ts
